Move rev2 to rev1
