+++ keyboards/crkbd/rev1/config.h
@@ -33,27 +33,21 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 /* key matrix size */
 // Rows are doubled-up
-#if CRKBD_ROWS == 3
-  #define MATRIX_ROWS 6
-  #define MATRIX_COLS 6
-  #define MATRIX_ROW_PINS { D7, E6, B4 }
-#elif CRKBD_ROWS == 4
+#if  CRKBD_ROWS == 4
   #define MATRIX_ROWS 8
-  #define MATRIX_COLS 6
-  #define MATRIX_ROW_PINS { D7, E6, B4, B5 }
+  #define MATRIX_COLS 7
+  #define MATRIX_ROW_PINS { D4, C6, D7, E6 }
 #elif CRKBD_ROWS == 5
   #define MATRIX_ROWS 10
-  #define MATRIX_COLS 6
-  #define MATRIX_ROW_PINS { D7, E6, B4, B5, D4 }
+  #define MATRIX_COLS 7
+  #define MATRIX_ROW_PINS { D4, C6, D7, E6, B4 }
 #else
-  #error "expected CRKBD_ROWS 3 or 4 or 5"
+  #error "expected CRKBD_ROWS 4 or 5"
 #endif
 
 // wiring of each half
-#define MATRIX_COL_PINS { F6, F7, B1, B3, B2, B6 }
-// #define MATRIX_COL_PINS { B6, B2, B3, B1, F7, F6 } //uncomment this line and comment line above if you need to reverse left-to-right key order
-
-#define CATERINA_BOOTLOADER
+#define MATRIX_COL_PINS { F4, F5, F6, F7, B1, B3, B2 }
+// #define MATRIX_COL_PINS { B2, B3, B1, F7, F6, F5, F4 } //uncomment this line and comment line above if you need to reverse left-to-right key order
 
 /* define if matrix has ghost */
 //#define MATRIX_HAS_GHOST
@@ -65,9 +59,9 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #define DEBOUNCING_DELAY 5
 
 /* Mechanical locking support. Use KC_LCAP, KC_LNUM or KC_LSCR instead in keymap */
-#define LOCKING_SUPPORT_ENABLE
+//#define LOCKING_SUPPORT_ENABLE
 /* Locking resynchronize hack */
-#define LOCKING_RESYNC_ENABLE
+//#define LOCKING_RESYNC_ENABLE
 
 /* key combination for command */
 #define IS_COMMAND() ( \
