+++ keyboards/crkbd/rev1/keymaps/default/config.h
@@ -23,15 +23,32 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "../../config.h"
 
-/* Use I2C or Serial, not both */
+/* Use I2C or Serial */
 
+#define USE_I2C
 #define USE_SERIAL
-// #define USE_I2C
+//#define USE_MATRIX_I2C
 
 /* Select hand configuration */
 
 #define MASTER_LEFT
-// #define _MASTER_RIGHT
+// #define MASTER_RIGHT
 // #define EE_HANDS
 
+#define SSD1306OLED
+
+#define USE_SERIAL_PD2
+
+#define PREVENT_STUCK_MODIFIERS
+#define TAPPING_FORCE_HOLD
+#define TAPPING_TERM 100
+
+
+#undef RGBLED_NUM
+#define RGBLIGHT_ANIMATIONS
+#define RGBLED_NUM 32
+#define RGBLIGHT_LIMIT_VAL 120
+#define RGBLIGHT_HUE_STEP 10
+#define RGBLIGHT_SAT_STEP 17
+#define RGBLIGHT_VAL_STEP 17
 #endif
