+++ keyboards/crkbd/rev1/rev1.c
@@ -1,5 +1,6 @@
 #include "crkbd.h"
 
+
 #ifdef AUDIO_ENABLE
     float tone_startup[][2] = SONG(STARTUP_SOUND);
     float tone_goodbye[][2] = SONG(GOODBYE_SOUND);
@@ -8,7 +9,7 @@
 #ifdef SSD1306OLED
 void led_set_kb(uint8_t usb_led) {
     // put your keyboard LED indicator (ex: Caps Lock LED) toggling code here
-    led_set_user(usb_led);
+    //led_set_user(usb_led);
 }
 #endif
 
@@ -19,21 +20,13 @@ void matrix_init_kb(void) {
         PLAY_SONG(tone_startup);
     #endif
 
-    // // green led on
-    // DDRD |= (1<<5);
-    // PORTD &= ~(1<<5);
-
-    // // orange led on
-    // DDRB |= (1<<0);
-    // PORTB &= ~(1<<0);
-
 	matrix_init_user();
 };
 
 void shutdown_kb(void) {
     #ifdef AUDIO_ENABLE
         PLAY_SONG(tone_goodbye);
-	_delay_ms(150);
-	stop_all_notes();
+      	_delay_ms(150);
+      	stop_all_notes();
     #endif
 }
