+++ keyboards/crkbd/rev1/rules.mk
@@ -1,4 +1,3 @@
 SRC += rev1/matrix.c \
-	   rev1/split_util.c
-
-BACKLIGHT_ENABLE = no
+	   rev1/split_util.c \
+	   ws2812.c
