+++ keyboards/crkbd/rev1/split_util.c
@@ -31,11 +31,7 @@ static void setup_handedness(void) {
 }
 
 static void keyboard_master_setup(void) {
-#ifdef USE_I2C
-#ifdef SSD1306OLED
-    matrix_master_OLED_init ();
-#endif
-#endif
+
 #ifdef USE_MATRIX_I2C
     i2c_master_init();
 #else
@@ -44,6 +40,7 @@ static void keyboard_master_setup(void) {
 }
 
 static void keyboard_slave_setup(void) {
+
 #ifdef USE_MATRIX_I2C
     i2c_slave_init(SLAVE_I2C_ADDRESS);
 #else
@@ -68,19 +65,7 @@ void split_keyboard_setup(void) {
    sei();
 }
 
-void keyboard_slave_loop(void) {
-   matrix_init();
-
-   while (1) {
-      matrix_slave_scan();
-   }
-}
-
 // this code runs before the usb and keyboard is initialized
 void matrix_setup(void) {
     split_keyboard_setup();
-
-    if (!has_usb()) {
-        keyboard_slave_loop();
-    }
 }
