+++ keyboards/crkbd/rev1/split_util.h
@@ -13,7 +13,6 @@ void matrix_slave_scan(void);
 
 void split_keyboard_setup(void);
 bool has_usb(void);
-void keyboard_slave_loop(void);
 
 void matrix_master_OLED_init (void);
