--- keyboards/crkbd/rev2/keymaps/crkbd/keymap.c
@@ -1,443 +1,0 @@
-#include "crkbd.h"
-#include "bootloader.h"
-#include "action_layer.h"
-#include "action_util.h"
-#include "eeconfig.h"
-#ifdef PROTOCOL_LUFA
-#include "lufa.h"
-#include "split_util.h"
-#endif
-#include "LUFA/Drivers/Peripheral/TWI.h"
-#ifdef AUDIO_ENABLE
-  #include "audio.h"
-#endif
-#ifdef SSD1306OLED
-  #include "ssd1306.h"
-#endif
-
-extern keymap_config_t keymap_config;
-
-#ifdef RGBLIGHT_ENABLE
-//Following line allows macro to read current RGB settings
-extern rgblight_config_t rgblight_config;
-#endif
-
-extern uint8_t is_master;
-
-// Each layer gets a name for readability, which is then used in the keymap matrix below.
-// The underscores don't mean anything - you can have a layer called STUFF or any other name.
-// Layer names don't all need to be of the same length, obviously, and you can also skip them
-// entirely and just use numbers.
-#define _QWERTY 0
-#define _LOWER 3
-#define _RAISE 4
-#define _ADJUST 16
-
-enum custom_keycodes {
-  QWERTY = SAFE_RANGE,
-  LOWER,
-  RAISE,
-  ADJUST,
-  BACKLIT,
-  EISU,
-  KANA,
-  RGBRST
-};
-
-enum macro_keycodes {
-  KC_SAMPLEMACRO,
-};
-
-
-// Fillers to make layering more clear
-#define _______ KC_TRNS
-#define XXXXXXX KC_NO
-//Macros
-#define M_SAMPLE M(KC_SAMPLEMACRO)
-
-#if CRKBD_ROWS == 5
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-  [_QWERTY] = KEYMAP( \
-  //,-----------------------------------------------------------------------.                          ,-----------------------------------------------------------------------.
-          KC_ESC,       KC_Q,       KC_W,       KC_E,       KC_R,       KC_T,                                  KC_Y,       KC_U,       KC_I,       KC_O,       KC_P,   KC_BSPC,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-        KC_LCTRL,       KC_A,       KC_S,       KC_D,       KC_F,       KC_G,                                  KC_H,       KC_J,       KC_K,       KC_L,    KC_SCLN,   KC_QUOT,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-         KC_LSFT,       KC_Z,       KC_X,       KC_C,       KC_V,       KC_B,                                  KC_N,       KC_M,    KC_COMM,     KC_DOT,    KC_SLSH,    KC_TAB,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------.  ,-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------|  |-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX \
-  //`-----------------------------------------------------------------------------------'  `-----------------------------------------------------------------------------------'
-  ),
-
-  [_LOWER] = KEYMAP( \
-  //,-----------------------------------------------------------------------.                          ,-----------------------------------------------------------------------.
-          KC_ESC,       KC_1,       KC_2,       KC_3,       KC_4,       KC_5,                                  KC_6,       KC_7,       KC_8,       KC_9,       KC_0,   KC_BSPC,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-        KC_LCTRL,      KC_F1,      KC_F2,      KC_F3,      KC_F4,      KC_F5,                                 KC_F6,      KC_F7,      KC_F8,      KC_F9,     KC_F10,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-         KC_LSFT,     KC_F11,     KC_F12,     KC_F13,     KC_F14,     KC_F15,                                KC_F16,     KC_F17,     KC_F18,     KC_F19,     KC_F20,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------.  ,-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------|  |-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX \
-  //`-----------------------------------------------------------------------------------'  `-----------------------------------------------------------------------------------'
-  ),
-
-  [_RAISE] = KEYMAP( \
-  //,-----------------------------------------------------------------------.                          ,-----------------------------------------------------------------------.
-          KC_ESC,    KC_EXLM,      KC_AT,    KC_HASH,     KC_DLR,    KC_PERC,                               KC_CIRC,    KC_AMPR,    KC_ASTR,   KC_LPRN,     KC_RPRN,   KC_BSPC,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-        KC_LCTRL,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,                               KC_MINS,     KC_EQL, S(KC_LBRC), S(KC_RBRC), S(KC_BSLS),    KC_GRV,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-         KC_LSFT,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,                            S(KC_MINS),  S(KC_EQL),    KC_LBRC,    KC_RBRC,    KC_BSLS, S(KC_GRV),\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------.  ,-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------|  |-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX \
-  //`-----------------------------------------------------------------------------------'  `-----------------------------------------------------------------------------------'
-  ),
-
-  [_ADJUST] = KEYMAP( \
-  //,-----------------------------------------------------------------------.                          ,-----------------------------------------------------------------------.
-           RESET,     RGBRST,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,                               XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-         RGB_TOG,    RGB_HUI,    RGB_SAI,    RGB_VAI,    XXXXXXX,    XXXXXXX,                               XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------|                          |-----------+-----------+-----------+-----------+-----------+-----------|
-        RGB_SMOD,    RGB_HUD,    RGB_SAD,    RGB_VAD,    XXXXXXX,    XXXXXXX,                               XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------.  ,-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX,\
-  //|-----------+-----------+-----------+-----------+-----------+-----------+-----------|  |-----------+-----------+-----------+-----------+-----------+-----------+-----------|
-         XXXXXXX,    XXXXXXX,    XXXXXXX,    KC_LGUI,      LOWER,     KC_SPC,    XXXXXXX,       XXXXXXX,     KC_ENT,      RAISE,    KC_RGUI,    XXXXXXX,    XXXXXXX,   XXXXXXX \
-  //`-----------------------------------------------------------------------------------'  `-----------------------------------------------------------------------------------'
-  )
-};
-
-#elif CRKBD_ROWS == 4
-#else
-#error "undefined keymaps"
-#endif
-
-#ifdef AUDIO_ENABLE
-
-float tone_qwerty[][2]     = SONG(QWERTY_SOUND);
-float tone_plover[][2]     = SONG(PLOVER_SOUND);
-float tone_plover_gb[][2]  = SONG(PLOVER_GOODBYE_SOUND);
-float music_scale[][2]     = SONG(MUSIC_SCALE_SOUND);
-#endif
-
-// define variables for reactive RGB
-bool TOG_STATUS = false;
-int RGB_current_mode;
-
-void persistent_default_layer_set(uint16_t default_layer) {
-  eeconfig_update_default_layer(default_layer);
-  default_layer_set(default_layer);
-}
-
-// Setting ADJUST layer RGB back to default
-void update_tri_layer_RGB(uint8_t layer1, uint8_t layer2, uint8_t layer3) {
-  if (IS_LAYER_ON(layer1) && IS_LAYER_ON(layer2)) {
-    #ifdef RGBLIGHT_ENABLE
-      //rgblight_mode(RGB_current_mode);
-    #endif
-    layer_on(layer3);
-  } else {
-    layer_off(layer3);
-  }
-}
-
-void matrix_init_user(void) {
-    #ifdef AUDIO_ENABLE
-        startup_user();
-    #endif
-    #ifdef RGBLIGHT_ENABLE
-      RGB_current_mode = rgblight_config.mode;
-    #endif
-    //SSD1306 OLED init, make sure to add #define SSD1306OLED in config.h
-    #ifdef SSD1306OLED
-        TWI_Init(TWI_BIT_PRESCALE_1, TWI_BITLENGTH_FROM_FREQ(1, 800000));
-        iota_gfx_init(!has_usb());   // turns on the display
-    #endif
-}
-
-
-#ifdef AUDIO_ENABLE
-
-void startup_user()
-{
-    _delay_ms(20); // gets rid of tick
-}
-
-void shutdown_user()
-{
-    _delay_ms(150);
-    stop_all_notes();
-}
-
-void music_on_user(void)
-{
-    music_scale_user();
-}
-
-void music_scale_user(void)
-{
-    PLAY_SONG(music_scale);
-}
-
-#endif
-
-
-//SSD1306 OLED update loop, make sure to add #define SSD1306OLED in config.h
-#ifdef SSD1306OLED
-
-void matrix_scan_user(void) {
-     iota_gfx_task();  // this is what updates the display continuously
-}
-
-void matrix_update(struct CharacterMatrix *dest,
-                          const struct CharacterMatrix *source) {
-  if (memcmp(dest->display, source->display, sizeof(dest->display))) {
-    memcpy(dest->display, source->display, sizeof(dest->display));
-    dest->dirty = true;
-  }
-}
-
-//assign the right code to your layers for OLED display
-#define L_BASE 0
-#define L_LOWER 8
-#define L_RAISE 16
-#define L_FNLAYER 64
-#define L_NUMLAY 128
-#define L_NLOWER 136
-#define L_NFNLAYER 192
-#define L_MOUSECURSOR 256
-#define L_ADJUST 65536
-#define L_ADJUST_TRI 65560
-
-static void render_logo(struct CharacterMatrix *matrix) {
-
-  static char logo[]={
-    0x80,0x81,0x82,0x83,0x84,0x85,0x86,0x87,0x88,0x89,0x8a,0x8b,0x8c,0x8d,0x8e,0x8f,0x90,0x91,0x92,0x93,0x94,
-    0xa0,0xa1,0xa2,0xa3,0xa4,0xa5,0xa6,0xa7,0xa8,0xa9,0xaa,0xab,0xac,0xad,0xae,0xaf,0xb0,0xb1,0xb2,0xb3,0xb4,
-    0xc0,0xc1,0xc2,0xc3,0xc4,0xc5,0xc6,0xc7,0xc8,0xc9,0xca,0xcb,0xcc,0xcd,0xce,0xcf,0xd0,0xd1,0xd2,0xd3,0xd4,
-    0};
-  matrix_write(matrix, logo);
-  //matrix_write_P(&matrix, PSTR(" Split keyboard kit"));
-}
-
-
-
-void render_status(struct CharacterMatrix *matrix, uint16_t keycode, keyrecord_t *record) {
-
-  // Render to mode icon
-  // static char logo[][2][3]={{{0x95,0x96,0},{0xb5,0xb6,0}},{{0x97,0x98,0},{0xb7,0xb8,0}}};
-  // if(keymap_config.swap_lalt_lgui==false){
-  //   matrix_write(matrix, logo[0][0]);
-  //   matrix_write_P(matrix, PSTR("\n"));
-  //   matrix_write(matrix, logo[0][1]);
-  // }else{
-  //   matrix_write(matrix, logo[1][0]);
-  //   matrix_write_P(matrix, PSTR("\n"));
-  //   matrix_write(matrix, logo[1][1]);
-  // }
-  //
-
-
-  // Render key
-  char key[40];
-  int led_matrix[10][7] = {
-    {  5,  4,  3,  2,  1,  0,  0},
-    {  6,  7,  8,  9, 10, 11,  0},
-    { 17, 16, 15, 14, 13, 12,  0},
-    { 18, 19, 20, 21, 22, 23, 24},
-    { 31, 30, 29, 28, 27, 26, 25},
-    { 37, 36, 35, 34, 33, 32,  0},
-    { 38, 39, 40, 41, 42, 43,  0},
-    { 49, 48, 47, 46, 45, 44,  0},
-    { 50, 51, 52, 53, 54, 55, 56},
-    { 63, 62, 61, 60, 59, 58, 57}
-  };
-
-  int led_num = led_matrix[record->event.key.row][record->event.key.col];
-  snprintf(key, sizeof(led), "%dx%d, k%2d l%2d",
-    record->event.key.row,
-    record->event.key.col,
-    keycode,
-    led_num
-  );
-  matrix_write(matrix, key);
-
-  // rgblight_disable();
-  // rgblight_enable();
-  // rgblight_setrgb_at(record->event.key.row*50,255-record->event.key.col*50,255, led_num);
-
-
-  // Define layers here, Have not worked out how to have text displayed for each layer. Copy down the number you see and add a case for it below
-  char buf[40];
-  snprintf(buf,sizeof(buf), "Undef-%ld", layer_state);
-  matrix_write_P(matrix, PSTR("\nLayer: "));
-    switch (layer_state) {
-        case L_BASE:
-           matrix_write_P(matrix, PSTR("Default"));
-           break;
-        case L_RAISE:
-           matrix_write_P(matrix, PSTR("Raise"));
-           break;
-        case L_LOWER:
-           matrix_write_P(matrix, PSTR("Lower"));
-           break;
-        case L_ADJUST:
-        case L_ADJUST_TRI:
-           matrix_write_P(matrix, PSTR("Adjust"));
-           break;
-        default:
-           matrix_write(matrix, buf);
-    }
-
-  // Host Keyboard LED Status
-  char led[40];
-    snprintf(led, sizeof(led), "\n%s  %s  %s",
-            (host_keyboard_leds() & (1<<USB_LED_NUM_LOCK)) ? "NUMLOCK" : "       ",
-            (host_keyboard_leds() & (1<<USB_LED_CAPS_LOCK)) ? "CAPS" : "    ",
-            (host_keyboard_leds() & (1<<USB_LED_SCROLL_LOCK)) ? "SCLK" : "    ");
-  matrix_write(matrix, led);
-}
-
-void iota_gfx_record_user(uint16_t keycode, keyrecord_t *record) {
-  struct CharacterMatrix matrix;
-
-  #if DEBUG_TO_SCREEN
-    if (debug_enable) {
-      return;
-    }
-  #endif
-
-    matrix_clear(&matrix);
-    if(is_master){
-      render_status(&matrix, keycode, record);
-    }else{
-      render_logo(&matrix);
-    }
-    matrix_update(&display, &matrix);
-}
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-  iota_gfx_record_user(keycode, record);
-
-  switch (keycode) {
-    case QWERTY:
-      if (record->event.pressed) {
-        #ifdef AUDIO_ENABLE
-          PLAY_SONG(tone_qwerty);
-        #endif
-        persistent_default_layer_set(1UL<<_QWERTY);
-      }
-      return false;
-      break;
-    case LOWER:
-      if (record->event.pressed) {
-          //not sure how to have keyboard check mode and set it to a variable, so my work around
-          //uses another variable that would be set to true after the first time a reactive key is pressed.
-        if (TOG_STATUS) { //TOG_STATUS checks is another reactive key currently pressed, only changes RGB mode if returns false
-        } else {
-          TOG_STATUS = !TOG_STATUS;
-          #ifdef RGBLIGHT_ENABLE
-            //rgblight_mode(16);
-          #endif
-        }
-        layer_on(_LOWER);
-        update_tri_layer_RGB(_LOWER, _RAISE, _ADJUST);
-      } else {
-        #ifdef RGBLIGHT_ENABLE
-          //rgblight_mode(RGB_current_mode);   // revert RGB to initial mode prior to RGB mode change
-        #endif
-        TOG_STATUS = false;
-        layer_off(_LOWER);
-        update_tri_layer_RGB(_LOWER, _RAISE, _ADJUST);
-      }
-      return false;
-      break;
-    case RAISE:
-      if (record->event.pressed) {
-        //not sure how to have keyboard check mode and set it to a variable, so my work around
-        //uses another variable that would be set to true after the first time a reactive key is pressed.
-        if (TOG_STATUS) { //TOG_STATUS checks is another reactive key currently pressed, only changes RGB mode if returns false
-        } else {
-          TOG_STATUS = !TOG_STATUS;
-          #ifdef RGBLIGHT_ENABLE
-            //rgblight_mode(15);
-          #endif
-        }
-        layer_on(_RAISE);
-        update_tri_layer_RGB(_LOWER, _RAISE, _ADJUST);
-      } else {
-        #ifdef RGBLIGHT_ENABLE
-          //rgblight_mode(RGB_current_mode);  // revert RGB to initial mode prior to RGB mode change
-        #endif
-        layer_off(_RAISE);
-        TOG_STATUS = false;
-        update_tri_layer_RGB(_LOWER, _RAISE, _ADJUST);
-      }
-      return false;
-      break;
-    case ADJUST:
-        if (record->event.pressed) {
-          layer_on(_ADJUST);
-        } else {
-          layer_off(_ADJUST);
-        }
-        return false;
-        break;
-      //led operations - RGB mode change now updates the RGB_current_mode to allow the right RGB mode to be set after reactive keys are released
-    case RGB_MOD:
-      #ifdef RGBLIGHT_ENABLE
-        if (record->event.pressed) {
-          rgblight_mode(RGB_current_mode);
-          rgblight_step();
-          RGB_current_mode = rgblight_config.mode;
-        }
-      #endif
-      return false;
-      break;
-    case EISU:
-      if (record->event.pressed) {
-        if(keymap_config.swap_lalt_lgui==false){
-          register_code(KC_LANG2);
-        }else{
-          SEND_STRING(SS_LALT("`"));
-        }
-      } else {
-        unregister_code(KC_LANG2);
-      }
-      return false;
-      break;
-    case KANA:
-      if (record->event.pressed) {
-        if(keymap_config.swap_lalt_lgui==false){
-          register_code(KC_LANG1);
-        }else{
-          SEND_STRING(SS_LALT("`"));
-        }
-      } else {
-        unregister_code(KC_LANG1);
-      }
-      return false;
-      break;
-    case RGBRST:
-      #ifdef RGBLIGHT_ENABLE
-        if (record->event.pressed) {
-          eeconfig_update_rgblight_default();
-          rgblight_enable();
-          RGB_current_mode = rgblight_config.mode;
-        }
-      #endif
-      break;
-  }
-  return true;
-}
-
-
-#endif
