--- keyboards/crkbd/rev2/rules.mk
@@ -1,3 +1,0 @@
-SRC += rev2/matrix.c \
-	   rev2/split_util.c \
-	   ws2812.c
