+++ keyboards/crkbd/rules.mk
@@ -70,4 +70,4 @@ SLEEP_LED_ENABLE = no    # Breathing sleep LED during USB suspend
 
 CUSTOM_MATRIX = yes
 
-DEFAULT_FOLDER = crkbd/rev2
+DEFAULT_FOLDER = crkbd/rev1
